bait/demo/requirements.txt
streamlit==1.41.1
pillow==9.5.0
opencv-python==4.10.0.84
ultralytics==8.3.19 
torch==2.5.1
scikit-learn==1.6.1
Unidecode==1.3.8
vietocr==0.3.12
pandas==2.2.3
openpyxl==3.1.5
transformers==4.36.1
datasets==2.21.0
tokenizers==0.15.2
seqeval==1.2.2
evaluate==0.4.3
pyvi==0.1.1
cupy-cuda11x==13.3.0
emoji==2.14.0
networkx==2.8.4
nltk==3.7
geocoder==1.38.1
python-dotenv==1.0.1
exif==1.6.1
